First initial commit for Four in a Row game.
ControlCentre handles the input from the user to create Players.
Player makes a move, and returns the move to control center.
Which in return sends it forward to GameState which will update the game.
Board class holds the logic for placing the pebble and checking for four in a row.